5_2_kafka/requirements.txt
certifi==2025.11.12
clickhouse-connect==0.10.0
kafka-python==2.3.0
lz4==4.4.5
psycopg2==2.9.11
pytz==2025.2
urllib3==2.5.0
zstandard==0.25.0
